Increase output precision of file1d
I noticed that sending floating point values through file1d leads to data files where the numbers have 6 digits or less.
I send raw data through file1d and then plot differences where the raw values are roughly in the range [1000:6000] and the differences are in the range [0:1] which leads to loss of precision. I did a quick hack to fix it so I can continue using it the way I do, it would probably be cleaner to somehow let the user choose the precision via an optional argument.